requirements.txt
future
emoji
requests
sqlalchemy
python-telegram-bot==13.2
psycopg2-binary
feedparser
toml
googletrans==4.0.0-rc1
pillow
alphabet-detector
flask
flask-restplus
Werkzeug==0.16.0
flask-httpauth
schedule
captcha